Player Movement and Camera
Used cinemachine to create the camera that works alongside the player. Player movement works independently from the camera, but going forwards will always to where the camera is facing.

Decided to go with a third person point of view as we wanted to have the player as a penguin - and we wouldn't be able to see it if it wasn't in the point of view.

Added one script, one camera, and a player subsititute.
The floor currently is not working!!